% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_digit.R
\name{table_digit}
\alias{table_digit}
\title{Round a numeric variable to a specified number of decimal places}
\usage{
table_digit(df, digit)
}
\arguments{
\item{df}{A dataframe for rounding a numeric variable}

\item{digit}{Number of decimal places}
}
\value{
A dataframe with only the numeric variables rounded
}
\description{
Round a numeric variable to a specified number of decimal places
}
