% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_xlsx.R
\name{export_xlsx}
\alias{export_xlsx}
\title{Output the results as a excel file}
\usage{
export_xlsx(output_list, file_path, sheet_name = "Report")
}
\arguments{
\item{output_list}{A list containing the results to be outputted}

\item{file_path}{The location and file name to save the file}

\item{sheet_name}{Sheet name}
}
\value{
An Excel file containing the output lists
}
\description{
Output the results as a excel file
}
